+++ tests/covalent_tests/test_covalent.py
@@ -28,6 +28,8 @@ def test_covalent_config():
     for executor in ct_config["executors"]:
         if "create_unique_workdir" in ct_config["executors"][executor]:
             assert ct_config["executors"][executor]["create_unique_workdir"] is True
+    if "slurm" in ct_config["executors"]:
+        assert ct_config["executors"]["slurm"].get("use_srun", True) is False
 
 
 @pytest.mark.skipif(