+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added this `CHANGELOG.md` file.
 - Added the `quacc.recipes.emt.parsl` module.
+- Added a CLI for `quacc` along with a `quacc config` option to configure Covalent appropriately.
 
 ### Changed